// src/interfaces/PersonaI.ts
export interface PersonaI {
    id?: number;
    tipoDocumento?: string;
    nroDocumento?: string;
    nombre?: string;
    apellidos?: string;
    genero?: string;
    fechaNacimiento?: string;
    celular?: string;
    email?: string;
    direccion?: string;
    status?: number;
    isDeleted?: number;
    createdAt?: string;
    updatedAt?: string;
}